+++ CHANGELOG.md
@@ -1,20 +1,29 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added Blaster application wrapper component [#173](https://github.com/raster-foundry/blasterjs/pull/173)
+
 ## [0.0.16] - 2019-03-20 [YANKED]
 
 ## [0.0.15] - 2019-03-20 [YANKED]
 
 ## [0.0.14] - 2019-03-11
+
 ### Added
+
 - Added chan for changelog management
 
 ## 0.0.14-alpha.1 - 2019-03-11
+
 ### Added
+
 - `Divider` component added to the core package [#94](https://github.com/raster-foundry/blasterjs/pull/94)
 - `Icon` component added to the core package [#103](https://github.com/raster-foundry/blasterjs/pull/103)
 - `Box` component added to the core package [#112](https://github.com/raster-foundry/blasterjs/pull/112)
@@ -27,15 +36,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `Table` components added to the core package [#132](https://github.com/raster-foundry/blasterjs/pull/132)
 
 ### Changed
+
 - Changed theme file `radius` to `radii` (breaking change for custom themes)
 - Normalized usage of styled-system across every component
 - Renamed `Link` to `A`.
 - Refactored theme system to support per-component themeing, replacing `defaultTheme.js` with a `theme` folder [#129](https://github.com/raster-foundry/blasterjs/pull/129)
 
 ### Removed
+
 - Removed reset.css.
 
 ### Fixed
+
 - Build script for docz documentation site and it now properly copies the reset.css file into the dist folder [#100](https://github.com/raster-foundry/blasterjs/pull/111)
 - CLI method `files` now properly filters by provided extension and now only `.js` files are used to generate the `index.common.js` file for each package [#120](https://github.com/raster-foundry/blasterjs/pull/120)